enviando

+++ src/main/java/br/com/motur/dealbackendservice/core/entrypoints/v1/request/VehicleDTO.java
@@ -78,5 +78,5 @@ public class VehicleDTO implements Serializable {
 
     @NotNull
     @Column(name = "provider_id")
-    private Set<Long> providerIds;
+    private Set<Integer> providerIds;
 }
